feat: refactor driver landing page
